backend/requirements.txt
# SkyGuard India - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Geospatial
geoalchemy2==0.14.3
shapely==2.0.2

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# HTTP Client (for Digital Sky API)
httpx==0.26.0
aiohttp==3.9.1

# Cryptography (for log chaining)
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
orjson==3.9.10

# XML Processing (for Permission Artifacts)
lxml==5.1.0
xmltodict==0.13.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.0
isort==5.13.2
mypy==1.8.0
